Refactor CreateGraphTransformer
